+++ src/main/resources/plugin.yml
@@ -3,17 +3,12 @@ version: '${version}'
 main: deponn.depmount.DepMount
 
 commands: #自作プラグインのコマンドを宣言
-  CreateMountain:
-    description: you can create mountain by using this command. /createmountain X_length Z_length opiton1(false for all or true for only air) opiton2(true for make boundary condition) option3(k of k-nn)
-    usage: please command /CreateMountain x(-150~150) z(-150~150) option1(0 or 1) option2(0 or 1) option3(0 or 1~)
-  UndoMountain:
-    description: you can undo CreateMountain command.
-    usage: only /UndoMountain
+  /mountain:
+    description: ラピスラズリブロックの点の補間により山を作成します。
+    usage: //mountain [-a|--all:空気以外も書き換え] [-b|--border:境界を作成] [-n|--num:補完の頂点数]
+    permission: mountain
 
 permissions:
-  CreateMountain:
-    description: op only
-    default: op
-  UndoMountain:
+  mountain:
     description: op only
     default: op
